Main Code
Still a work in progess.
Using a webscraping method, the program goes to the most acessed lyrics and can click on each of the 1000 most listened songs and collect lyrics, clean the html stuff and count each instance of a word.
Still need to save it to a csv or DataFrame to work with the collected data better